+++ xknx/io/ip_secure.py
@@ -426,12 +426,6 @@ class SecureGroup(UDPTransport, _IPSecureTransportLayer):
             self.secure_timer.handle_timer_notify(knxipframe.body)
             return
         if isinstance(knxipframe.body, SecureWrapper):
-            if not self.secure_timer.validate_secure_wrapper(knxipframe.body):
-                ip_secure_logger.warning(
-                    "Discarding SecureWrapper with invalid timer value: %s",
-                    knxipframe,
-                )
-                return
             if not self.secure_timer.timer_authenticated:
                 ip_secure_logger.debug(
                     "Discarding received SecureWrapper before timer synchronisazion finished: %s",
@@ -439,7 +433,7 @@ class SecureGroup(UDPTransport, _IPSecureTransportLayer):
                 )
                 return
             try:
-                knxipframe = self.decrypt_frame(knxipframe)
+                decrypted_knxipframe = self.decrypt_frame(knxipframe)
             except KNXSecureValidationError as err:
                 ip_secure_logger.warning("Could not decrypt KNXIPFrame: %s", err)
                 # Frame shall be discarded
@@ -450,8 +444,14 @@ class SecureGroup(UDPTransport, _IPSecureTransportLayer):
                     couldnotparseknxip.description,
                 )
                 return
-            knx_logger.debug("Decrypted frame: %s", knxipframe)
-        super().handle_knxipframe(knxipframe, source)
+            if not self.secure_timer.validate_secure_wrapper(knxipframe.body):
+                ip_secure_logger.warning(
+                    "Discarding SecureWrapper with invalid timer value: %s",
+                    knxipframe,
+                )
+                return
+            knx_logger.debug("Decrypted frame: %s", decrypted_knxipframe)
+        super().handle_knxipframe(decrypted_knxipframe, source)
 
     def send(self, knxipframe: KNXIPFrame, addr: tuple[str, int] | None = None) -> None:
         """Send KNXIPFrame to socket. `addr` is ignored on TCP."""
@@ -476,6 +476,8 @@ class SecureSequenceTimer:
 
     According to AN159 v06 KNXnet-IP Secure AS §2.2.2.3 Timer synchronizing
     """
+
+    TIMER_NOTIFY_HEADER = bytes.fromhex("06 10 09 55 00 24")
 
     min_delay_time_keeper_periodic_notify: Final = 10  # pylint: disable=invalid-name
     min_delay_time_keeper_update_notify: Final = 0.1  # pylint: disable=invalid-name
@@ -605,10 +607,9 @@ class SecureSequenceTimer:
         b_0 = timer_bytes + serial_number + _message_tag + b"\x00\x00"
         # TODO: get header data and total_length from TimerNotify class
         #       or handle mac calculation in TimerNotify class directly
-        timer_notify_header = bytes.fromhex("06 10 09 55 00 24")
         mac_cbc = calculate_message_authentication_code_cbc(
             key=self._backbone_key,
-            additional_data=timer_notify_header,
+            additional_data=self.TIMER_NOTIFY_HEADER,
             block_0=b_0,
         )
         c_0 = timer_bytes + serial_number + _message_tag + b"\xff\x00"
@@ -628,6 +629,34 @@ class SecureSequenceTimer:
             ),
             None,
         )
+
+    def verify_timer_notify_mac(self, timer_notify: TimerNotify) -> None:
+        """Verify MAC of timer notify."""
+        timer_bytes = timer_notify.timer_value.to_bytes(6, "big")
+        b_0 = (
+            timer_bytes
+            + timer_notify.serial_number
+            + timer_notify.message_tag
+            + b"\x00\x00"
+        )
+        c_0 = (
+            timer_bytes
+            + timer_notify.serial_number
+            + timer_notify.message_tag
+            + b"\xff\x00"
+        )
+        _, mac_tr = decrypt_ctr(
+            key=self._backbone_key,
+            counter_0=c_0,
+            mac=timer_notify.message_authentication_code,
+        )
+        mac_cbc = calculate_message_authentication_code_cbc(
+            key=self._backbone_key,
+            additional_data=self.TIMER_NOTIFY_HEADER,
+            block_0=b_0,
+        )
+        if mac_cbc != mac_tr:
+            raise KNXSecureValidationError("MAC verification failed")
 
     async def synchronize(self) -> None:
         """Synchronize timer with remote time keeper."""
@@ -685,6 +714,13 @@ class SecureSequenceTimer:
     def handle_timer_notify(self, timer_notify: TimerNotify) -> None:
         """Handle received TimerNotify frame."""
         local_timer_value = self.current_timer_value()
+        try:
+            self.verify_timer_notify_mac(timer_notify)
+        except KNXSecureValidationError:
+            ip_secure_logger.warning(
+                "Discarding TimerNotify with invalid MAC: %s", timer_notify
+            )
+            return
         received_timer_value = timer_notify.timer_value
         # §2.2.2.3.2.5 Events: E11
         if (
@@ -693,7 +729,7 @@ class SecureSequenceTimer:
             and timer_notify.message_tag == self._expected_notify_handler[0]
         ):
             fut = self._expected_notify_handler[1]
-            fut.set_result(timer_notify.timer_value)
+            fut.set_result(received_timer_value)
             return
         # §2.2.2.3.2.5 Events: E1 - E4
         if received_timer_value > local_timer_value: