+++ docs/changelog.md
@@ -8,6 +8,11 @@ nav_order: 2
 
 # Unreleased changes
 
+### IP Secure
+
+- SecureRouting: verify MAC of received TimerNotify frames.
+- SecureRouting: verify and handle timer value of received SecureWrapper frames after verification of MAC.
+
 ### Internals
 
 - Parse T_Data_Broadcast TPCI. Forward these telegrams to the Management class.